Finish allocator rewrite (it works)

+++ include/kernel/memory/kPageAllocator.h
@@ -3,7 +3,7 @@
 #include "libraries/hardware/debug.h"
 #include "kernel/memory/kSlabAllocator.h"
 #include "kernel/memory/kMemoryPage.h"
-#include "kernel/memory/vmm.h"
+#include "kernel/memory/mmu.h"
 
 #define MAX_PAGES 256
 
@@ -11,7 +11,7 @@ static char pageBitmap[MAX_PAGES / 8] = {};
 
 void PageAllocator_Init();
 int64_t PageAllocator_AllocPages(size_t pageCnt);
-void PageAllocator_FreePages(uint32_t page, size_t count);
+void PageAllocator_FreePages(uint64_t page, size_t count);
 
 uint8_t PageAllocator_GetPageStatus(uint64_t page);
 uint64_t PageAllocator_UsedPagesCount();
